adding an uploaded field to cookbook json #116
This changes adds an uploaded field to the cookbook JSON that gets
pushed. We would like this field since we sometimes have folks that forget
to bump the cookbook versions and upload the cookbook to the chef server.
In this case, because there is no change to the JSON, no commit is created
in the git repo so no event is fired off from a github webhook.

This ensures that for every knife cookbook upload, chef guard will be
able to commit something to git.