We should not touch the content of the files
This also fixes #107 where we found corrupted binaries in cookbooks that Chef-Guard pushed into the supermarket.

At the same time it reverts the fix for issue #72. So to accomodate for issues mentioned in that issue, we extended the error message so it points out CRLF vs LF differences could cause the diff.

To solve these types of compare errors you probably need to add this line to your git config:

```
Checkout Unix-style, commit Unix-style line endings
```